+++ build-system/compile/post-closure-babel.js
@@ -16,25 +16,11 @@
 'use strict';
 const argv = require('minimist')(process.argv.slice(2));
 const babel = require('@babel/core');
-const fs = require('fs-extra');
 const path = require('path');
 const remapping = require('@ampproject/remapping');
 const terser = require('terser');
 const through = require('through2');
 const {debug, CompilationLifecycles} = require('./debug-compilation-lifecycle');
-
-/**
- * Given a filepath, return the sourcemap.
- *
- * @param {string} file
- * @return {string|null}
- */
-function loadSourceMap(file) {
-  if (file.startsWith('dist')) {
-    return fs.readFile(`${file}.map`);
-  }
-  return null;
-}
 
 /**
  * Minify passed string.
